# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShopWave E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user's cart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear entire cart
      responses:
        "204":
          description: Cart cleared
  /cart/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ─── WISHLIST ──────────────────────────────────────────────────────────────
...
  /addresses:
    get:
      operationId: listAddresses
      tags: [addresses]
      summary: List user's saved addresses
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: createAddress
      tags: [addresses]
      summary: Add a new address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Created address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /addresses/{addressId}:
    put:
      operationId: updateAddress
      tags: [addresses]
      summary: Update address
      parameters:
        - name: addressId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    delete:
      operationId: deleteAddress
      tags: [addresses]
      summary: Delete address
      parameters:
        - name: addressId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── USERS / PROFILE ───────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{userId}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban or unban a user
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/orders:
    get:
      operationId: listAllOrders
      tags: [admin]
      summary: List all platform orders (admin)
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
components:
  schemas:
...
    CartItem:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        productTitle: { type: string }
        productImage: { type: string }
        productPrice: { type: number }
        finalPrice: { type: number }
        discountPercent: { type: number }
        variant: { type: ["string", "null"] }
        quantity: { type: integer }
        stock: { type: integer }
      required: [id, productId, productTitle, productImage, productPrice, finalPrice, discountPercent, quantity, stock]
    Cart:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/CartItem" } }
        subtotal: { type: number }
        discountAmount: { type: number }
        total: { type: number }
        itemCount: { type: integer }
      required: [items, subtotal, discountAmount, total, itemCount]
    CartItemInput:
      type: object
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        variant: { type: ["string", "null"] }
      required: [productId, quantity]
    CartItemUpdate:
      type: object
      properties:
        quantity: { type: integer }
      required: [quantity]

    # ─── WISHLIST ─────────────────────────────────────────────────────────────
...
    OrderItem:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        productTitle: { type: string }
        productImage: { type: string }
        variant: { type: ["string", "null"] }
        quantity: { type: integer }
        price: { type: number }
      required: [id, productId, productTitle, productImage, quantity, price]
    OrderTimeline:
      type: object
      properties:
        status: { type: string }
        message: { type: string }
        timestamp: { type: string }
      required: [status, message, timestamp]
    Order:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        items: { type: array, items: { $ref: "#/components/schemas/OrderItem" } }
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, return_requested, returned]
        paymentMethod: { type: string }
        paymentStatus: { type: string }
        subtotal: { type: number }
        discountAmount: { type: number }
        gstAmount: { type: number }
        shippingFee: { type: number }
        total: { type: number }
        couponCode: { type: ["string", "null"] }
        trackingNumber: { type: ["string", "null"] }
        estimatedDelivery: { type: ["string", "null"] }
        timeline: { type: array, items: { $ref: "#/components/schemas/OrderTimeline" } }
        shippingAddress: { $ref: "#/components/schemas/Address" }
        createdAt: { type: string }
      required: [id, userId, items, status, paymentMethod, paymentStatus, subtotal, discountAmount, gstAmount, shippingFee, total, timeline, createdAt]
    OrderList:
      type: object
      properties:
        orders: { type: array, items: { $ref: "#/components/schemas/Order" } }
        total: { type: integer }
        page: { type: integer }
        totalPages: { type: integer }
      required: [orders, total, page, totalPages]
...
    Address:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        name: { type: string }
        phone: { type: string }
        line1: { type: string }
        line2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        country: { type: string }
        isDefault: { type: boolean }
      required: [id, userId, name, phone, line1, city, state, pincode, country, isDefault]
    AddressInput:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        line1: { type: string }
        line2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        country: { type: string }
        isDefault: { type: boolean }
      required: [name, phone, line1, city, state, pincode, country]

    # ─── USER ─────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id: { type: integer }
        clerkId: { type: string }
        email: { type: string }
        name: { type: string }
        avatarUrl: { type: ["string", "null"] }
        role:
          type: string
          enum: [customer, seller, admin]
        isBanned: { type: boolean }
        createdAt: { type: string }
      required: [id, clerkId, email, name, role, isBanned, createdAt]
...
    BanUserInput:
      type: object
      properties:
        banned: { type: boolean }
        reason: { type: string }
      required: [banned]

    # ─── COUPON ───────────────────────────────────────────────────────────────
...