# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopWave E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from './address';
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';
import type { OrderTimeline } from './orderTimeline';

export interface Order {
  id: number;
  userId: number;
  items: OrderItem[];
  status: OrderStatus;
  paymentMethod: string;
  paymentStatus: string;
  subtotal: number;
  discountAmount: number;
  gstAmount: number;
  shippingFee: number;
  total: number;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  trackingNumber?: string | null;
  /** @nullable */
  estimatedDelivery?: string | null;
  timeline: OrderTimeline[];
  shippingAddress?: Address;
  createdAt: string;
}