# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopWave E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  isBanned: boolean;
  createdAt: string;
}